Make move chips rollable
This adapts the rollable helper to work on ratings rather than direct
dice rolls which allows us to very easily do moves without having to
look anything up on the actor and also apply +1 forwards once that's
implemented.

I switched the .click handler to a .on handler since they're deprecated
in modern jQuery.